+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Some examples, more below in the actual changelog (newer entries are more likely
 
 ### Added
 * core/v1: helper methods Tag, Untag and ListTags (#122, @marioreggiori)
+* lbaas/v1: ACL and Rule API bindings added (#142, @toothstone & @marioreggiori)
 
 ### Changed
 * moved pkg/api.GetObjectIdentifier and related errors to pkg/api/types (#144, @LittleFox94)